Add files via upload
EVM stands for Electronics Voting Machine. It’s a lot different from the traditional ballot paper system both in terms of mechanism and performance. Its USP is its simple user interface. Even a person who never got a chance to go to schools can use it without much difficulty. The front panel shows all the candidates standing for the election along with their party symbols. There is a button corresponding to each of the candidates. To cast a vote just press the button beside to the candidate. A successful vote is indicated by a green light and a short beep. There is a dedicated counter for each of the candidate ,which is placed inside. With each vote the counter corresponding to the candidate increases and is displayed through a LCD screen. This arrangement is kept under lock. After the election’s over the polling officer can open the lock and view the votes and declare the result.